// src/types/driver.ts
// src/types/driver.ts
export interface DriverSelectDTO {
  id: number;
  nome: string;
  cpf: string;
  celular: string;
}

export interface DriverDTO {
  codigoId: number;
  nomeMot: string;
  cpfMot: string;
  rgMot: string;
  estadoMot: string;
  dataEmiss: string | null;
  dataNasc: string | null;
  cidadeNas: string;
  estadoNas: string;
  cnhMot: string;
  regCnh: string;
  ufEmissor: string;
  data1Habi: string | null;
  catCnh: string;
  vencCnh: string | null;
  maeMot: string;
  paiMot: string;
  endMot: string;
  cepMot: string;
  bairroMot: string;
  cidAtual: string;
  ufAtual: string;
  telRes: string;
  cont1Mot: string;
  telCom: string;
  cont2Mot: string;
  ref1Mot: string;
  cont3Mot: string;
  ref2Mot: string;
  cont4Mot: string;
  placaVeic: string;
  cidVeic: string;
  estadoVei: string;
  renavan: string;
  marcaVeic: string;
  modeloVei: string;
  chassi: string;
  anoVeic: string;
  corVeic: string;
  tipoVeic: string;
  numResMo: string;
  munMot: string;
  codMot: string;
  venctoBon: string | null;
  modelo: string;
  propNom: string;
  propCpf: string;
  propRg: string;
  propCep: string;
  propEnd: string;
  propNum: string;
  propBai: string;
  propMun: string;
  propUf: string;
  frota: string;
  carCpf: string;
  carRg: string;
  carrNom: string;
  carEnd: string;
  carNum: string;
  carMun: string;
  carCep: string;
  carUf: string;
  rntc: string;
  cnhSeg: string;
  dtaInc: string | null;
  placaCar: string;
  bcoMot: string;
  ctaMot: string;
  agencia: string;
  nomFav: string;
  bcoOp: string;
  anttVcto: string | null;
  antt: string;
  qtdEixo: number | null;
  altExt: number | null;
  compExt: number | null;
  largExt: number | null;
  rastreado: string;
  qtdPal: number | null;
  pesoL: number | null;
  regiaoDes: string;
  nivelApp: string;
  senhaApp: string;
  apiFirebase: string;
  bloqueioA: string;
  bloqueioApp: string;
  datau: string | null;
  horau: string;
  emailMot: string;
  pisMot: string;
  dtEmiCnh: string | null;
  rgOrgEmi: string;
  tpMot: string;
  tpVinMot: string;
  numPagbem: string;
  hrInc: string;
  uCodInc: string;
  nCodInc: string;
  motInss: string;
  motTpPix: string;
  motChPix: string;
  motNCvc: string;
  motPAvc: number | null;
  bcoCod: string;
  status: string;
}